Fix inability to access NodePort in particular case
When a Service NodePort and an Egress CRD has the same backend Pod, accessing
to the NodePort Service may fail in particular cases. Assume that the backend
Pod is on Node A and the Egress's external IP is on Node B. If an external
client (not any K8s Node) accesses the NodePort through IP of Node A where
the backend Pod is running, the access will fail. The root cause is that the
reply packets of NodePort is incorrectly matched by the flow installed by Egress
which is used to match the packets sourced from local Pods and destined for
tunneling to Node B. This PR fixes the issue by loading FromGatewayCTMark to
Service connections from Antrea gateway, then the reply packets will be forced
back to Antrea gateway.

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>