Support limiting number of Egress IPs that can be assigned to a Node
The current implementation of Egress uses 8 bits of pkt mark to indicate
which SNAT IP the egress traffic should use, which means at most 255
Egress IPs can be assigned to a single Node. Besides, some Node network
may restrict the number of secondary IPs that can be assigned to a Node,
e.g. EKS.

This patch adds a configuration option maxEgressIPsPerNode and makes it
default to 255. When a Node cannot accommodate more Egress IPs, the next
Node will be tried, until an available Node is found or all Nodes are
tried.

Signed-off-by: Quan Tian <qtian@vmware.com>